+++ FMI3_to_VDM/src/main/java/fmi2vdm/elements/NumberOfEventIndicators.java
@@ -34,7 +34,7 @@ import org.xml.sax.Locator;
 
 public class NumberOfEventIndicators extends Element
 {
-	private Integer[] dependencies;
+	private Long[] dependencies;
 
 	public NumberOfEventIndicators(Attributes attributes, Locator locator)
 	{
